requirements.txt
aniso8601==6.0.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
flask-restplus==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
marshmallow==2.19.2
passlib==1.7.1
pg8000==1.13.1
pkg-resources==0.0.0
pyrsistent==0.15.1
python-dotenv==0.10.1
pytz==2019.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.3
Werkzeug==0.15.2
